Simphony-CAPS-Backup - Downloadable Exe File
BackupBuddy is a handy utility designed to automatically back up Oracle Simphony Client Workstation databases, including "checkposting" and "datastore," which are hosted on Microsoft SQL Server (MSSQL). The tool creates backups and logs, storing them in the "C:\BackupBuddy\" folder by default. It is designed to work with Simphony versions 2.7, 2.9, 2.10, 18.x, and 19.x, offering a simple and reliable solution for data protection.

- Automatic backups for Oracle Simphony Client Workstation databases hosted on MSSQL.
- Backs up "checkposting" and "datastore" databases by default.
- Saves backups and logs in the "C:\BackupBuddy\" folder.
- Backups will be saved with the file extension of .sg
- Developed with default "sa" user and password "#1mymicros."
- Change the sa user name and password in EMC Property Parameters (Under admin credentials and restart the CAPS/Terminals)
- Supports Simphony versions 2.7, 2.9, 2.10, 18.x, and 19.x.